# backend/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon Demo API
  description: >
    OpenAPI 3.0 规范，覆盖 Hackathon Demo 后端的全部 REST 接口，方便前后端联调与测试。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
  /api/v1/events:
    post:
      tags: [Events]
      summary: 创建活动
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: 请求校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Events]
      summary: 获取活动列表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '500':
          description: 服务端错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/events/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Events]
      summary: 获取单个活动
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Events]
      summary: 更新活动
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Events]
      summary: 删除活动
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/events/{id}/stage:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [Events]
      summary: 更新活动阶段
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStageRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sponsors:
    post:
      tags: [Sponsors]
      summary: 创建赞助商
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSponsorRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Sponsors]
      summary: 获取赞助商列表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sponsor'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sponsors/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Sponsors]
      summary: 获取赞助商
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Sponsors]
      summary: 更新赞助商
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSponsorRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Sponsors]
      summary: 删除赞助商
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sponsors/address/{address}:
    parameters:
      - name: address
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Sponsors]
      summary: 按地址查询赞助商
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sponsorships:
    post:
      tags: [Sponsorships]
      summary: 创建赞助
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSponsorshipRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsorship'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sponsorships/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Sponsorships]
      summary: 获取赞助详情
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsorship'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sponsorships/event/{eventId}:
    parameters:
      - $ref: '#/components/parameters/EventIdPathParam'
    get:
      tags: [Sponsorships]
      summary: 获取活动赞助列表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sponsorship'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sponsorships/{id}/approve:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [Sponsorships]
      summary: 审批赞助
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerActionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsorship'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/sponsorships/{id}/reject:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [Sponsorships]
      summary: 拒绝赞助
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerActionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsorship'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/sponsorships/{id}/deposit:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [Sponsorships]
      summary: 更新存款状态
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositStatusRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsorship'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/funding-pools:
    post:
      tags: [FundingPools]
      summary: 创建奖池
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFundingPoolRequest'
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingPool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [FundingPools]
      summary: 奖池列表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FundingPool'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/funding-pools/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [FundingPools]
      summary: 获取奖池
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingPool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [FundingPools]
      summary: 更新奖池金额
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFundingPoolRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingPool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/funding-pools/event/{eventId}:
    parameters:
      - $ref: '#/components/parameters/EventIdPathParam'
    get:
      tags: [FundingPools]
      summary: 按活动查询奖池
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingPool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/funding-pools/event/{eventId}/lock:
    parameters:
      - $ref: '#/components/parameters/EventIdPathParam'
    patch:
      tags: [FundingPools]
      summary: 设置锁定时间
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLockedUntilRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingPool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/funding-pools/event/{eventId}/distribute:
    parameters:
      - $ref: '#/components/parameters/EventIdPathParam'
    patch:
      tags: [FundingPools]
      summary: 标记奖池已分发
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingPool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/teams:
    post:
      tags: [Teams]
      summary: 创建团队
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Teams]
      summary: 团队列表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/teams/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Teams]
      summary: 获取团队
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Teams]
      summary: 更新团队
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Teams]
      summary: 删除团队
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/teams/leader/{address}:
    parameters:
      - $ref: '#/components/parameters/AddressPathParam'
    get:
      tags: [Teams]
      summary: 按队长地址查询团队
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/teams/member/{address}:
    parameters:
      - $ref: '#/components/parameters/AddressPathParam'
    get:
      tags: [Teams]
      summary: 按成员地址查询团队
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/teams/{id}/members:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    post:
      tags: [Teams]
      summary: 添加成员
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/teams/{id}/members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
      - name: memberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags: [Teams]
      summary: 移除成员
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/teams/{id}/approve:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [Teams]
      summary: 审批团队
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerActionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/teams/{id}/reject:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [Teams]
      summary: 拒绝团队
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerActionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/registrations:
    post:
      tags: [Registrations]
      summary: 提交团队报名
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegistrationRequest'
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/registrations/event/{eventId}:
    parameters:
      - $ref: '#/components/parameters/EventIdPathParam'
    get:
      tags: [Registrations]
      summary: 获取活动报名列表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Registration'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/registrations/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Registrations]
      summary: 获取报名
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Registrations]
      summary: 删除报名
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/registrations/{id}/approve:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [Registrations]
      summary: 审批报名
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerActionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/registrations/{id}/reject:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [Registrations]
      summary: 拒绝报名
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerActionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/registrations/{id}/sbt:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [Registrations]
      summary: 更新 SBT 铸造状态
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSBTStatusRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/check-ins:
    post:
      tags: [CheckIns]
      summary: 签到校验
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInRequest'
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIn'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [CheckIns]
      summary: （保留）支持未来扩展
      responses:
        '200':
          description: OK
  /api/v1/check-ins/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [CheckIns]
      summary: 获取签到记录
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [CheckIns]
      summary: 删除签到
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/check-ins/{id}/tx:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    patch:
      tags: [CheckIns]
      summary: 更新签到交易哈希
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTxHashRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/check-ins/event/{eventId}:
    parameters:
      - $ref: '#/components/parameters/EventIdPathParam'
    get:
      tags: [CheckIns]
      summary: 活动签到列表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckIn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/check-ins/event/{eventId}/qrcode:
    parameters:
      - $ref: '#/components/parameters/EventIdPathParam'
    get:
      tags: [CheckIns]
      summary: 生成签到二维码
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInQRCodeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/check-ins/event/{eventId}/count:
    parameters:
      - $ref: '#/components/parameters/EventIdPathParam'
    get:
      tags: [CheckIns]
      summary: 获取签到人数
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    EventIdPathParam:
      name: eventId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: 请求错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务端错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      example:
        error: resource not found
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    UpdateStageRequest:
      type: object
      required: [stage]
      properties:
        stage:
          $ref: '#/components/schemas/EventStage'
...
    CreateSponsorRequest:
      type: object
      required: [name, address]
      properties:
        name:
          type: string
        description:
          type: string
        logo_url:
          type: string
        website_url:
          type: string
        address:
          type: string
...